third_party/blink/web_tests/clipboard/async-clipboard/async-clipboard-read-types-rejection.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>Async clipboard selective format read</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/clipboard-write.js"></script>
<script src="resources/user-activation.js"></script>
<script>
// Tests clipboard.read() parameter validation with invalid types parameter.
// Validates TypeError rejection for non-array types: numbers, strings, objects, functions.
// Ensures proper type checking before clipboard access.
promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: 42}));
}, 'types: number should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: "text/plain"}));
}, 'types: string should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: document.createElement("div")}));
}, 'types: DOM element should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: {}}));
}, 'types: object should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: () => {}}));
}, 'types: function should throw TypeError');
</script>
